#4153 Refactoring by moving method getBilltoDefaultLocationIdByUser to IBPartnerDAO
#4153 Serial Letter Feature
