+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/bpartner/service/IBPartnerDAO.java
@@ -29,6 +29,8 @@ import java.util.Set;
 
 import org.adempiere.bpartner.BPartnerId;
 import org.adempiere.bpartner.BPartnerType;
+import org.adempiere.location.LocationId;
+import org.adempiere.user.User;
 import org.adempiere.util.ISingletonService;
 import org.compiere.model.I_C_BP_Relation;
 import org.compiere.model.I_C_BPartner;
@@ -230,4 +232,6 @@ public interface IBPartnerDAO extends ISingletonService
 	I_AD_User retrieveContact(Properties ctx, int bpartnerId, boolean isSOTrx, String trxName);
 
 	Map<BPartnerId, Integer> retrieveAllDiscountSchemaIdsIndexedByBPartnerId(BPartnerType bpartnerType);
+
+	LocationId getBilltoDefaultLocationIdByUser(@NonNull final User user);
 }
