+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/bpartner/service/impl/BPartnerDAO.java
@@ -41,12 +41,17 @@ import org.adempiere.bpartner.BPartnerId;
 import org.adempiere.bpartner.BPartnerType;
 import org.adempiere.bpartner.service.IBPartnerDAO;
 import org.adempiere.bpartner.service.OrgHasNoBPartnerLinkException;
+import org.adempiere.location.Location;
+import org.adempiere.location.LocationId;
+import org.adempiere.location.LocationRepository;
 import org.adempiere.model.InterfaceWrapperHelper;
+import org.adempiere.user.User;
 import org.adempiere.util.Check;
 import org.adempiere.util.GuavaCollectors;
 import org.adempiere.util.NumberUtils;
 import org.adempiere.util.Services;
 import org.adempiere.util.proxy.Cached;
+import org.compiere.Adempiere;
 import org.compiere.model.IQuery;
 import org.compiere.model.I_C_BP_Relation;
 import org.compiere.model.I_C_BPartner;
@@ -618,5 +623,23 @@ public class BPartnerDAO implements IBPartnerDAO
 			default:
 				return null;
 		}
+	}
+
+	@Override
+	public LocationId getBilltoDefaultLocationIdByUser(@NonNull final User user)
+	{
+		if (user.getBpartnerId() == null)
+		{
+			return null;
+		}
+		final LocationRepository repo = Adempiere.getBean(LocationRepository.class);
+		return retrieveBPartnerLocations(user.getBpartnerId())
+				.stream()
+				.filter(I_C_BPartner_Location::isBillToDefault)
+				.map(I_C_BPartner_Location::getC_Location)
+				.map(locationRecord -> repo.ofRecord(locationRecord))
+				.findFirst()
+				.map(Location::getId)
+				.orElse(null);
 	}
 }
